apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentprovision-api"
version = "0.1.0"
description = "FastAPI backend for AgentProvision.com"
authors = [
  { name = "AgentProvision", email = "eng@agentprovision.com" }
]
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "sqlalchemy>=2.0.25",
  "asyncpg>=0.29.0",
  "alembic>=1.13.1",
  "pydantic-settings>=2.1.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "redis>=5.0.1",
  "celery>=5.3.6",
  "python-dotenv>=1.0.0",
  "httpx>=0.26.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.2.0",
  "pytest>=7.4.4",
  "pytest-asyncio>=0.23.3"
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["app"]
